debug-requirements.txt
Django==1.10.5
django-redis==4.7.0
django-simple-captcha==0.5.3
django-solo==1.1.2
psycopg2==2.6.2
pudb==2016.2